# Allow controlling time flow for EmbeddedEventLoop (#12459)