--- a/common/src/main/java/io/netty5/util/concurrent/SingleThreadEventExecutor.java
+++ b/common/src/main/java/io/netty5/util/concurrent/SingleThreadEventExecutor.java
@@ -250,7 +250,7 @@ public class SingleThreadEventExecutor extends AbstractScheduledEventExecutor im
     }
 
     private boolean fetchFromScheduledTaskQueue() {
-        long nanoTime = AbstractScheduledEventExecutor.nanoTime();
+        long nanoTime = getCurrentTimeNanos();
         RunnableScheduledFuture<?> scheduledTask  = pollScheduledTask(nanoTime);
         while (scheduledTask != null) {
             if (!taskQueue.offer(scheduledTask)) {
@@ -379,6 +379,7 @@ public class SingleThreadEventExecutor extends AbstractScheduledEventExecutor im
      */
     protected final long delayNanos(long currentTimeNanos) {
         assert inEventLoop();
+        currentTimeNanos -= START_TIME;
         RunnableScheduledFuture<?> scheduledTask = peekScheduledTask();
         if (scheduledTask == null) {
             return SCHEDULE_PURGE_INTERVAL;
@@ -388,7 +389,7 @@ public class SingleThreadEventExecutor extends AbstractScheduledEventExecutor im
     }
 
     /**
-     * Returns the absolute point in time (relative to {@link #nanoTime()}) at which the the next
+     * Returns the absolute point in time (relative to {@link #getCurrentTimeNanos()} ()}) at which the next
      * closest scheduled task should run.
      *
      * This method must be called from the {@link EventExecutor} thread.
@@ -397,7 +398,7 @@ public class SingleThreadEventExecutor extends AbstractScheduledEventExecutor im
         assert inEventLoop();
         RunnableScheduledFuture<?> scheduledTask = peekScheduledTask();
         if (scheduledTask == null) {
-            return nanoTime() + SCHEDULE_PURGE_INTERVAL;
+            return getCurrentTimeNanos() + SCHEDULE_PURGE_INTERVAL;
         }
         return scheduledTask.deadlineNanos();
     }
@@ -412,7 +413,7 @@ public class SingleThreadEventExecutor extends AbstractScheduledEventExecutor im
      */
     protected final void updateLastExecutionTime() {
         assert inEventLoop();
-        lastExecutionTime = nanoTime();
+        lastExecutionTime = getCurrentTimeNanos();
     }
 
     /**
@@ -599,7 +600,7 @@ public class SingleThreadEventExecutor extends AbstractScheduledEventExecutor im
         cancelScheduledTasks();
 
         if (gracefulShutdownStartTime == 0) {
-            gracefulShutdownStartTime = nanoTime();
+            gracefulShutdownStartTime = getCurrentTimeNanos();
         }
 
         if (runAllTasks() || runShutdownHooks()) {
@@ -618,7 +619,7 @@ public class SingleThreadEventExecutor extends AbstractScheduledEventExecutor im
             return false;
         }
 
-        final long nanoTime = nanoTime();
+        final long nanoTime = getCurrentTimeNanos();
 
         if (isShutdown() || nanoTime - gracefulShutdownStartTime > gracefulShutdownTimeout) {
             return true;