--- a/transport/src/main/java/io/netty5/channel/embedded/EmbeddedEventLoop.java
+++ b/transport/src/main/java/io/netty5/channel/embedded/EmbeddedEventLoop.java
@@ -28,6 +28,22 @@ import java.util.concurrent.TimeUnit;
 import static java.util.Objects.requireNonNull;
 
 final class EmbeddedEventLoop extends AbstractScheduledEventExecutor implements EventLoop {
+    /**
+     * When time is not {@link #timeFrozen frozen}, the base time to subtract from {@link System#nanoTime()}. When time
+     * is frozen, this variable is unused.
+     *
+     * Initialized to {@link #initialNanoTime()} so that until one of the time mutator methods is called,
+     * {@link #getCurrentTimeNanos()} matches the default behavior.
+     */
+    private long startTime = initialNanoTime();
+    /**
+     * When time is frozen, the timestamp returned by {@link #getCurrentTimeNanos()}. When unfrozen, this is unused.
+     */
+    private long frozenTimestamp;
+    /**
+     * Whether time is currently frozen.
+     */
+    private boolean timeFrozen;
 
     private final Queue<Runnable> tasks = new ArrayDeque<>(2);
     boolean running;
@@ -91,7 +107,7 @@ final class EmbeddedEventLoop extends AbstractScheduledEventExecutor implements 
     }
 
     long runScheduledTasks() {
-        long time = AbstractScheduledEventExecutor.nanoTime();
+        long time = getCurrentTimeNanos();
         boolean wasRunning = running;
         try {
             for (;;) {
@@ -120,6 +136,40 @@ final class EmbeddedEventLoop extends AbstractScheduledEventExecutor implements 
             cancelScheduledTasks();
         } finally {
             running = false;
+        }
+    }
+
+    @Override
+    protected long getCurrentTimeNanos() {
+        if (timeFrozen) {
+            return frozenTimestamp;
+        }
+        return System.nanoTime() - startTime;
+    }
+
+    void advanceTimeBy(long nanos) {
+        if (timeFrozen) {
+            frozenTimestamp += nanos;
+        } else {
+            // startTime is subtracted from nanoTime, so increasing the startTime will advance getCurrentTimeNanos
+            startTime -= nanos;
+        }
+    }
+
+    void freezeTime() {
+        if (!timeFrozen) {
+            frozenTimestamp = getCurrentTimeNanos();
+            timeFrozen = true;
+        }
+    }
+
+    void unfreezeTime() {
+        if (timeFrozen) {
+            // we want getCurrentTimeNanos to continue right where frozenTimestamp left off:
+            // getCurrentTimeNanos = nanoTime - startTime = frozenTimestamp
+            // then solve for startTime
+            startTime = System.nanoTime() - frozenTimestamp;
+            timeFrozen = false;
         }
     }